backend: deleted star-system router/controller/repository in favor of gameplay and return the star system in the player view
