+++ backend/src/api/gameplay/gameplay.repository.ts
@@ -1,9 +1,10 @@
-import { Assert, type Logger, Result } from "@guillaume-docquier/tools-ts"
+import { Assert, type Logger, Range, Result } from "@guillaume-docquier/tools-ts"
 import { and, eq } from "drizzle-orm"
 import type { AccountId } from "#api/accounts/AccountId.ts"
 import type { GameId } from "#api/shared/GameId.ts"
 import type { PlayerId } from "#api/shared/PlayerId.ts"
-import type { NewStarSystemModel, NewSectorModel } from "#api/star-systems/starSystems.repository.ts"
+import { toCoordinates } from "#api/star-systems/Coordinates.ts"
+import type { Transaction } from "#lib/db/createDb.ts"
 import type { GamePlayerActionType } from "#lib/db/gameplay/gamePlayerActionType.ts"
 import { ResourceType } from "#lib/db/gameplay/gameResources.ts"
 import { PostgresRepository } from "#lib/db/PostgresRepository.ts"
@@ -22,6 +23,7 @@ import {
   ticksTable,
 } from "#lib/db/schema.ts"
 import { couldNot } from "#lib/errors.ts"
+import type { BodyType } from "#lib/star-systems/BodyType.ts"
 
 type NewGameStateRow = typeof gameStatesTable.$inferInsert
 type GameStateRow = typeof gameStatesTable.$inferSelect
@@ -38,9 +40,109 @@ export type PlayerViewModel = {
   playerId: PlayerId
   tick: number
   nextTickAt: Date
+  starSystem: StarSystemModel
   resources: {
     money: number
   }
+}
+
+const RANGE_NUMERIC_TYPES = ["float", "integer"] as const
+const RANGE_MAX_BOUND_TYPES = ["inclusive", "exclusive"] as const
+
+type StarSystemRow = typeof starSystemsTable.$inferSelect
+type OrbitRow = typeof orbitsTable.$inferSelect
+type SectorRow = typeof sectorsTable.$inferSelect
+type BodyRow = typeof bodiesTable.$inferSelect
+type MovementEdgeRow = typeof movementEdgesTable.$inferSelect
+
+export type NewStarSystemModel = {
+  orbits: NewOrbitModel[]
+  sectors: NewSectorModel[]
+  bodies: NewBodyModel[]
+  movementNodes: NewMovementNodeModel[]
+  movementEdges: NewMovementEdgeModel[]
+}
+
+export type NewOrbitModel = {
+  id: string
+  orbitNumber: number
+}
+
+export type NewSectorModel = {
+  id: string
+  orbitId: string
+  sectorNumber: number
+  angleRange: Range
+  movementNodeId: string
+}
+
+export type NewBodyModel = {
+  id: string
+  sectorId: string
+  bodyNumber: number
+  bodyType: BodyType
+  name: string
+  movementNodeId: string
+}
+
+export type NewMovementNodeModel = {
+  id: string
+}
+
+export type NewMovementEdgeModel = {
+  fromNodeId: string
+  toNodeId: string
+  weight: number
+}
+
+export type StarSystemModel = {
+  /**
+   * Star system as a tree
+   */
+  orbits: OrbitModel[]
+  /**
+   * Movement edges by movement node id
+   */
+  movementEdges: Record<string, MovementEdgeModel[]>
+}
+
+export type OrbitModel = {
+  id: string
+  number: number
+  coordinates: string
+  sectors: SectorModel[]
+}
+
+export type SectorModel = {
+  id: string
+  number: number
+  coordinates: string
+  angleRange: Range
+  bodies: BodyModel[]
+  movementNodeId: string
+}
+
+export type BodyModel = {
+  id: string
+  number: number
+  coordinates: string
+  name: string
+  type: BodyType
+  movementNodeId: string
+}
+
+export type MovementEdgeModel = {
+  fromNodeId: string
+  toNodeId: string
+  weight: number
+}
+
+type StarSystemAggregatedRows = {
+  starSystem: StarSystemRow
+  orbits: OrbitRow[]
+  sectors: SectorRow[]
+  bodies: BodyRow[]
+  movementEdges: MovementEdgeRow[]
 }
 
 export type StartGameModel = {
@@ -59,6 +161,9 @@ export type StartGameModel = {
   >
 }
 
+/**
+ * As the repository grows in size because of the models, we'll want to start extracting models/queries in their own files to keep it manageable.
+ */
 export class GameplayRepository extends PostgresRepository {
   private readonly logger: Logger
 
@@ -91,7 +196,6 @@ export class GameplayRepository extends PostgresRepository {
       tick: gameState.tick,
       scheduledFor: startGameModel.nextTickAt,
     }
-    Assert.isTrue(startGameModel.starSystem.gameId === startGameModel.gameId)
 
     const startResult = await Result.tryCatch(
       db.transaction(async (tx) => {
@@ -99,7 +203,7 @@ export class GameplayRepository extends PostgresRepository {
         await tx.insert(gameStatesTable).values(gameState)
         await tx.insert(resourcesTable).values(playerResources)
         await tx.insert(ticksTable).values(gameTick)
-        await storeStarSystem(startGameModel.starSystem, tx)
+        await insertStarSystem({ gameId: startGameModel.gameId, starSystem: startGameModel.starSystem }, tx)
       }),
     )
 
@@ -138,6 +242,16 @@ export class GameplayRepository extends PostgresRepository {
           return undefined
         }
 
+        const starSystems = await tx.select().from(starSystemsTable).where(eq(starSystemsTable.gameId, gameId))
+        Assert.isTrue(starSystems.length === 1)
+        const starSystem = starSystems[0]
+        Assert.isDefined(starSystem)
+
+        const orbits = await tx.select().from(orbitsTable).where(eq(orbitsTable.gameId, gameId))
+        const sectors = await tx.select().from(sectorsTable).where(eq(sectorsTable.gameId, gameId))
+        const bodies = await tx.select().from(bodiesTable).where(eq(bodiesTable.gameId, gameId))
+        const movementEdges = await tx.select().from(movementEdgesTable).where(eq(movementEdgesTable.gameId, gameId))
+
         const playerResources = await tx
           .select()
           .from(resourcesTable)
@@ -148,6 +262,7 @@ export class GameplayRepository extends PostgresRepository {
         return {
           ...gameState,
           playerId,
+          starSystem: toStarSystemModel({ starSystem, orbits, sectors, bodies, movementEdges }),
           resources: {
             money: money.amount,
           },
@@ -284,34 +399,37 @@ export class GameplayRepository extends PostgresRepository {
   }
 }
 
-async function storeStarSystem(starSystem: NewStarSystemModel, db: PostgresRepository["db"]): Promise<void> {
-  const withGameId = createWithGameId(starSystem.gameId)
+async function insertStarSystem(
+  { gameId, starSystem }: { gameId: GameId; starSystem: NewStarSystemModel },
+  tx: Transaction,
+): Promise<void> {
+  const withGameId = createWithGameId(gameId)
 
-  await db.insert(starSystemsTable).values(withGameId({}))
+  await tx.insert(starSystemsTable).values(withGameId({}))
 
   const movementNodes = starSystem.movementNodes.map(withGameId)
   if (movementNodes.length > 0) {
-    await db.insert(movementNodesTable).values(movementNodes)
+    await tx.insert(movementNodesTable).values(movementNodes)
   }
 
   const movementEdges = starSystem.movementEdges.map(withGameId)
   if (movementEdges.length > 0) {
-    await db.insert(movementEdgesTable).values(movementEdges)
+    await tx.insert(movementEdgesTable).values(movementEdges)
   }
 
   const orbits = starSystem.orbits.map(withGameId)
   if (orbits.length > 0) {
-    await db.insert(orbitsTable).values(orbits)
+    await tx.insert(orbitsTable).values(orbits)
   }
 
   const sectors = starSystem.sectors.map(toNewSectorRow).map(withGameId)
   if (sectors.length > 0) {
-    await db.insert(sectorsTable).values(sectors)
+    await tx.insert(sectorsTable).values(sectors)
   }
 
   const bodies = starSystem.bodies.map(withGameId)
   if (bodies.length > 0) {
-    await db.insert(bodiesTable).values(bodies)
+    await tx.insert(bodiesTable).values(bodies)
   }
 }
 
@@ -328,8 +446,55 @@ function toNewSectorRow(newSector: NewSectorModel): NewSectorRow {
   }
 }
 
-function createWithGameId(
-  gameId: NewStarSystemModel["gameId"],
-): <T extends Record<string, unknown>>(data: T) => T & Pick<NewStarSystemModel, "gameId"> {
+function createWithGameId(gameId: GameId): <T extends Record<string, unknown>>(data: T) => T & { gameId: GameId } {
   return (data) => ({ ...data, gameId })
+}
+
+export function toStarSystemModel(starSystemRows: StarSystemAggregatedRows): StarSystemModel {
+  const bodiesBySectorId = Map.groupBy(starSystemRows.bodies, ({ sectorId }) => sectorId)
+  const sectorsByOrbitId = Map.groupBy(starSystemRows.sectors, ({ orbitId }) => orbitId)
+
+  // It's a bit monstrous, but it's localized and does exactly what it need to
+  return {
+    orbits: starSystemRows.orbits.map((orbit) => ({
+      id: orbit.id,
+      number: orbit.orbitNumber,
+      coordinates: toCoordinates({ orbitNumber: orbit.orbitNumber }),
+      sectors: (sectorsByOrbitId.get(orbit.id) ?? []).map((sector) => ({
+        id: sector.id,
+        number: sector.sectorNumber,
+        coordinates: toCoordinates({ orbitNumber: orbit.orbitNumber, sectorNumber: sector.sectorNumber }),
+        angleRange: toSectorAngleRange(sector),
+        movementNodeId: sector.movementNodeId,
+        bodies: (bodiesBySectorId.get(sector.id) ?? []).map((body) => ({
+          id: body.id,
+          number: body.bodyNumber,
+          coordinates: toCoordinates({ orbitNumber: orbit.orbitNumber, sectorNumber: sector.sectorNumber, bodyNumber: body.bodyNumber }),
+          name: body.name,
+          type: body.bodyType,
+          movementNodeId: body.movementNodeId,
+        })),
+      })),
+    })),
+    movementEdges: toMovementEdgesByFromNodeId(starSystemRows.movementEdges),
+  }
+}
+
+function toSectorAngleRange(sector: SectorRow): Range {
+  Assert.isOneOf(RANGE_NUMERIC_TYPES, sector.angleNumericType, "sector.angleNumericType")
+  Assert.isOneOf(RANGE_MAX_BOUND_TYPES, sector.angleMaxBoundType, "sector.angleMaxBoundType")
+
+  return Range.create({
+    numericType: sector.angleNumericType,
+    maxBoundType: sector.angleMaxBoundType,
+    min: sector.startAngleDegrees,
+    max: sector.endAngleDegrees,
+  })
+}
+
+function toMovementEdgesByFromNodeId(edges: MovementEdgeRow[]): StarSystemModel["movementEdges"] {
+  // We cast because Object.groupBy returns a Partial<Record<string, T>>, which makes TypeScript think
+  // That T could be undefined because of Partial
+  // Kinda strange
+  return Object.groupBy(edges, ({ fromNodeId }) => fromNodeId) as StarSystemModel["movementEdges"]
 }
